+++ src/main/java/model/Model.java
@@ -3,22 +3,22 @@ package model;
 import java.io.Serializable;
 
 public class Model implements Serializable {
-    private Long id;
+    private int id;
 
     public Model() {
 
     }
 
 
-    public Model(Long id){
+    public Model(int id){
         this.id = id;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
